refactor - moved extractTags fucntion to utils
